feat(dotcom): store comment bodies as rich text in the Postgres projection
The comment record's body was already rich text end-to-end through the room
and R2 lane, but the Postgres/Zero projection flattened it to a plaintext
column. Store the TLRichText JSON as-is (JSONB) so the projected copy
preserves the authoritative representation:

- migration 039: comment.text (VARCHAR) -> comment.body (JSONB), backfilling
  existing plaintext rows as single-paragraph rich text docs
- Zero schema: comment.body is a json() column
- DO projection passes record.body through instead of flattening
- basic UI flattens to plaintext at display time only (shared
  richTextToPlaintext util); rich text input/rendering UI forthcoming
